Added ItemStack#setMaterial and information about automatic refresh

+++ src/main/java/net/minestom/server/inventory/PlayerInventory.java
@@ -212,7 +212,7 @@ public class PlayerInventory implements InventoryModifier, InventoryClickHandler
      *
      * @param slot the slot to refresh
      */
-    public void refreshSlot(int slot) {
+    public void refreshSlot(short slot) {
         sendSlotRefresh((short) convertToPacketSlot(slot), getItemStack(slot));
     }
